+++ package.json
@@ -19,12 +19,13 @@
   "homepage": "https://github.com/Termanty/MovesApp-backend#readme",
   "devDependencies": {
     "@types/express": "^4.17.13",
-    "@types/node": "^17.0.13",
-    "ts-node": "^10.4.0",
+    "@types/node": "^17.0.17",
+    "ts-node": "^10.5.0",
     "typescript": "^4.5.5"
   },
   "dependencies": {
     "cors": "^2.8.5",
+    "dotenv": "^16.0.0",
     "express": "^4.17.2",
     "mariadb": "^2.5.5"
   }
